pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = "backend django server"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.23.3"
djangorestframework = "^3.14.0"
Markdown = "^3.4.1"
django-filter = "^22.1"
Django = "^4.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
